+++ packages/grid-export/src/features/grid-excel-export/transformers/__tests__/ConfigTransformer.test.ts
@@ -2,7 +2,7 @@ import {
   testConfig,
   testConfigWithGroups,
   testItems,
-} from "../__mocks__/TestData";
+} from "../__mocks__/ExcelExportTestData";
 import { createZipcelxConfig } from "../ConfigTransformer";
 
 describe("ConfigTransformer", () => {