+++ packages/grid-export/src/features/grid-excel-export/transformers/__tests__/HeaderTransformer.test.ts
@@ -2,7 +2,7 @@ import {
   testConfig,
   testGroupConfigs,
   testGroupConfigsWithGroups,
-} from "../__mocks__/TestData";
+} from "../__mocks__/ExcelExportTestData";
 import {
   transformGroupHeaders,
   transformTableHeaders,